some bugs in lesson modesl and dto classes fixed and create new lesson method added in the lesson controller

+++ src/main/java/com/schoolmanagement/schoolmanangemenet/repository/ILessonRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface ILessonRepository extends CrudRepository<Lesson, Integer> {
+public interface ILessonRepository extends CrudRepository<Lesson, Long> {
     Optional<Lesson> findByName(String lessonName);
 }
