+++ src/main/java/com/schoolmanagement/schoolmanangemenet/service/interfaces/ILessonService.java
@@ -13,5 +13,5 @@ public interface ILessonService {
 
     Boolean update (UpdateLessonDTO updateLessonDTO);
 
-    Boolean delete (Integer lessonId);
+    Boolean delete (Long lessonId);
 }
